requirements.txt
django-paypal==0.5.0
